Comments out before_action CAUSING DEPLOYED BUG
The before action (generated by rails) for all three controllers was
designed to set_<model_name> on :show, :update, and :destroy. However,
with the way I designed this API, the show method is called against
normal conventions. The :id being passed is that of the user, or some
other model. So it's doing what I want it to do with 'show'
(despite how un-rails that may be) but it's also calling the before
action to locate that model by 'id' -> BUT the 'id' I'm passing is that
of user, or date_master or something else.

THE REASON this does not cause problems locally is that I've generated
so much data that it's still finding those resources, even though I'm
not doing anything with it. However, the deployed database with throw
an error as soon as it attempts that before action and comes up empty.